Lint warnings and refactor
